requirements.txt
pandas
numpy
mlflow
scikit-learn
matplotlib
seaborn
dvc
streamlit
python-box
ensure
-e .